+++ packages/core/pipeline/src/core/run-history.ts
@@ -203,6 +203,7 @@ export class RunnableCollection {
     if (runnable?.status) {
       runnable.status.status = ResultType.none;
       runnable.status.result = ResultType.none;
+      runnable.status.output = {};
       runnable.status.inputHash = "";
       // @ts-ignore
       runnable.input = {};