+++ packages/ui/certd-client/src/components/vip-button/api.ts
@@ -4,7 +4,7 @@ export async function doActive(form: any) {
   return await request({
     url: "/sys/plus/active",
     method: "post",
-    data: form
+    data: form,
   });
 }
 
@@ -12,6 +12,6 @@ export async function getVipTrial() {
   return await request({
     url: "/sys/plus/getVipTrial",
     method: "post",
-    data: {}
+    data: {},
   });
 }