+++ packages/ui/certd-server/src/controller/user/mine/email-controller.ts
@@ -17,4 +17,25 @@ export class EmailController extends BaseController {
     await this.emailService.test(userId, receiver);
     return this.ok({});
   }
+
+  @Post('/list', { summary: Constants.per.authOnly })
+  public async list() {
+    const userId = super.getUserId();
+    const res = await this.emailService.list(userId);
+    return this.ok(res);
+  }
+
+  @Post('/add', { summary: Constants.per.authOnly })
+  public async add(@Body('email') email) {
+    const userId = super.getUserId();
+    await this.emailService.add(userId, email);
+    return this.ok({});
+  }
+
+  @Post('/delete', { summary: Constants.per.authOnly })
+  public async delete(@Body('email') email) {
+    const userId = super.getUserId();
+    await this.emailService.delete(userId, email);
+    return this.ok({});
+  }
 }